requirements.txt
# ──────────────────────────────────────────────
# Japanese OCR Translation Assistant
# requirements.txt
# Python 3.10+ recommended
# ──────────────────────────────────────────────

# ── UI ──────────────────────────────────────
PyQt6>=6.5.0

# ── Screen Capture ───────────────────────────
mss>=9.0.1
Pillow>=10.0.0

# ── OCR ──────────────────────────────────────
paddlepaddle>=2.5.2          # CPU-only PaddlePaddle
paddleocr>=2.7.0             # PaddleOCR with Japanese support

# ── NLP / Morphological Analysis ─────────────
fugashi>=1.3.0               # Python MeCab wrapper
unidic-lite>=1.0.8           # Lightweight UniDic dictionary

# ── Translation ───────────────────────────────
transformers>=4.35.0         # HuggingFace Transformers (MarianMT)
sentencepiece>=0.1.99        # Required by MarianMT tokenizer
torch>=2.0.0                 # PyTorch (CPU; for GPU add +cu118 index)
numpy>=1.24.0

# ── Utilities ────────────────────────────────
requests>=2.31.0             # For any HTTP needs
tqdm>=4.66.0                 # Progress bars during download
